handle post errors

+++ src/blog/blog.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Post } from './post.schema';
@@ -15,7 +15,9 @@ export class BlogService {
     if (!id.match(/^[0-9a-fA-F]{24}$/)) {
       throw new BadRequestException('Invalid id');
     }
-    return this.postModel.findById(id).exec();
+    const post = await this.postModel.findById(id).exec();
+    if (!post) throw new NotFoundException();
+    return post;
   }
 
   async createPost(
@@ -30,12 +32,22 @@ export class BlogService {
   }
 
   async updatePost(id: string, title: string, content: string): Promise<Post> {
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+      throw new BadRequestException('Invalid id');
+    }
+    const post = await this.postModel.findById(id).exec();
+    if (!post) throw new NotFoundException();
     return this.postModel
       .findByIdAndUpdate(id, { title, content }, { new: true })
       .exec();
   }
 
   async deletePost(id: string): Promise<Post> {
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+      throw new BadRequestException('Invalid id');
+    }
+    const post = await this.postModel.findById(id).exec();
+    if (!post) throw new NotFoundException();
     return this.postModel.findByIdAndDelete(id).exec();
   }
 }
